# specs/005-user-org-service/contracts/user-org-service.openapi.yaml
openapi: 3.1.0
info:
  title: User & Organization Service API
  version: 0.1.0
  description: >
    Contracts for interactive admin API, authentication flows, policy decisions,
    budget enforcement, declarative reconciliation controls, and audit exports.
servers:
  - url: https://user-org-service.{env}.platform.internal
    description: Cluster-scoped service endpoint (dev/staging/prod)
  - url: http://localhost:8081
    description: Local development (admin API)
...
security:
  - bearerAuth: []
  - m2mToken: []
paths:
...
  /v1/orgs/{orgId}/service-accounts:
    post:
      tags: [ServiceAccounts]
      summary: Create a service account
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceAccountRequest'
      responses:
        '201':
          description: Service account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceAccount'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Refresh token with dedicated scope
    m2mToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Machine-to-machine token issued to downstream services
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
        format: slug
...
  schemas:
...
    CreateServiceAccountRequest:
      type: object
      required: [name, roles]
      properties:
        name:
          type: string
        description:
          type: string
        roles:
          type: array
          items:
            type: string
    ServiceAccount:
      type: object
      required: [serviceAccountId, name, status]
      properties:
        serviceAccountId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, disabled, deleted]
        roles:
          type: array
          items:
            type: string
...